+++ config/lerk.rb
@@ -55,8 +55,6 @@ Hello there!
 You have reached the rate limit of Hive 2 queries.
 Please wait a few seconds before you issue another query. :)
 EOF
-
-      ENABLE_EMBEDS = ['true', 'yes', '1'].include?(ENV.fetch('HIVE_ENABLE_EMBEDS', 'true'))
     end
 
     module Observatory