+++ docker-entrypoint.sh
@@ -8,18 +8,18 @@ echo "APPLICATION_ENV=$APPLICATION_ENV"
 case "$1" in
     application)
         echo "Starting application server..."
-        exec puma -C config/puma.rb
+        exec bundle exec puma -C config/puma.rb
         ;;
     bot)
         echo "Starting discord bot..."
-        exec ruby bin/lerk
+        exec bundle exec ruby bin/lerk
         ;;
     migrate)
         echo "Applying database migrations"
-        exec rake db:migrate
+        exec bundle exec rake db:migrate
         ;;
     *)
         echo "Don't know what to do with $1"
-        echo "Valid commands: bot, migrate"
+        echo "Valid commands: application, bot, migrate"
         exit 1
 esac